cleaning up the JSON marshaller, moving test data, and adding a benchmark or two
